Rename record_indexer/by_collection.py -> record_indexer/create_collection_index.py
